+++ test/unit/lib/device/handler/RequestHandler_test.js
@@ -1,33 +1,68 @@
 'use strict';
 
-const expect = require('chai').expect;
-const Request = require('../../../../../../../lib/device/express/routes/handler/request');
+const chai = require('chai');
+chai.use(require('sinon-chai'));
+const expect = chai.expect;
+const sinon = require('sinon');
+const RequestHandler = require('../../../../../lib/device/handler/RequestHandler');
 
-describe('./lib/device/express/routes/handler/request.js', function() {
+describe('./lib/device/handler/index.js', function() {
+  let db;
+  let requestHandler;
+
+  beforeEach(function() {
+    db = {
+      search: sinon.stub(),
+      getDevice: sinon.stub()
+    };
+    requestHandler = new RequestHandler(db);
+  });
+
+  it('should handle searchDevice', function() {
+    // GIVEN
+    const query = 'NEEO Brain';
+
+    // WHEN
+    requestHandler.searchDevice(query);
+
+    // THEN
+    expect(db.search).to.have.been.calledWith(query);
+  });
+
+  it('should handle getDevice', function() {
+    // GIVEN
+    const id = 'NEEO Brain';
+
+    // WHEN
+    requestHandler.getDevice(id);
+
+    // THEN
+    expect(db.getDevice).to.have.been.calledWith(id);
+  });
 
   it('should fail to call discover without controller', function() {
-    return Request.discover()
+    return requestHandler.discover()
       .catch((error) => {
         expect(error.message).to.equal('INVALID_DISCOVER_PARAMETER');
       });
   });
 
   it('should fail to call discover with invalid controller', function() {
-    return Request.discover({ controller: 123 })
+    return requestHandler.discover({ controller: 123 })
       .catch((error) => {
         expect(error.message).to.equal('CONTROLLER_IS_NOT_A_FUNCTION');
       });
   });
 
   it('should fail to call handleGet without parameter', function() {
-    return Request.handleGet()
+    return requestHandler.handleGet()
       .catch((error) => {
         expect(error.message).to.equal('INVALID_GET_PARAMETER');
       });
   });
 
   it('should fail to call handleGet with invalid componenttype', function() {
-    return Request.handleGet({
+    return requestHandler.handleGet({
         handler: {
           componenttype: 'foo',
           controller: {}
@@ -39,7 +74,7 @@ describe('./lib/device/express/routes/handler/request.js', function() {
   });
 
   it('should fail to call handleGet without controller', function() {
-    return Request.handleGet({
+    return requestHandler.handleGet({
         handler: {
           componenttype: 'switch',
           controller: {}
@@ -51,14 +86,14 @@ describe('./lib/device/express/routes/handler/request.js', function() {
   });
 
   it('should fail to call handleSet without parameter', function() {
-    return Request.handleSet()
+    return requestHandler.handleSet()
       .catch((error) => {
         expect(error.message).to.equal('INVALID_SET_PARAMETER');
       });
   });
 
   it('should fail to call handleSet without controller', function() {
-    return Request.handleSet({
+    return requestHandler.handleSet({
         handler: {
           componenttype: 'switch'
         },
@@ -69,7 +104,7 @@ describe('./lib/device/express/routes/handler/request.js', function() {
   });
 
   it('should fail to call handleSet without controller', function() {
-    return Request.handleSet({
+    return requestHandler.handleSet({
         handler: {
           componenttype: 'switch',
           controller: {}